old and new tests - will rationalise later
